refactor: extract tooltip into separate file
